Add QLineEditModel and QLineEditEventHelper
The idea was to extract the canvas independent bits from QLineControl
and make it suitable for using inside Qml. The drawing code, input
method and completion code was removed. Completion we should look
later with more care since it's both a model and UI thing.

The overall API of the class wasn't changed much, just added some
Q_PROPERTY and Q_SLOTS.

The code for QTextLayout is special in the sense that it is both a
model and non-model issue. For now, we kept the text layout inside the
model, and we can interact with it by setting the font and
getting/setting the cursorX position.

A helper class for dealing with events was also added and a small qml
example of usage.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>