Revamp of QDateTimeModel
Cleanup of QDateTimeModel in order to remove stuff that was from QDateTimeEdit
and also revamped it in order to use the convenient methods from QDateTime,
QDate and QTime.

This model provides a wide range of features and properties, allowing one to
create a lot of date/time related views. This model can be expanded in order
to allow some other complex features like holidays.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>