Add Style
This is the basic class in the component styling system discussed
during the week.  The idea is that a Style will not "paint stuff"
(imperative painting) like QStyle, but instead it will give a
component a subtree of other canvas items.  Our implementation will
consider that the basic 'styleable' components are QGraphicsObjects,
but the idea can be applied in a more general way.

Besides the common interface, Style also has a basic implementation
for the populate() method, that look in a table for 'populators'
previously registered using component name and style name as key.

The basic platform Styles can benefit from this default implementation
by just changing the style name, and providing new populators.

The 'populator' mechanism serve as a substitute for the
'switch-on-type' implementation of populate(). One can easily override
populate() and use 'switch-on-type' if needed.

We are experimenting with the model idea, so we are explicitly passing
model to the populate() method for now.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>