Add TestStyle
This is an example of a regular style that makes use of the default
Style implementation.  Besides setting the style name in the style
class, it also provides a baseclass for populators implementing this
style (which will set the correct style name to use when registering
populators).

The superclass as convenience may provide primitives and access
style-specific parameters to easy the life of populators'
implementors. In the example, it provides access to a "rectangle"
primitive that is part of private implementation of the style.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>