Add QmlStyle
Creates a different style that overrides the default Style::populate()
method and instead of looking at the 'populator' table, look for a
qml-file in a given directory, and produces a QDeclarativeItem from
that. It falls back to looking for a 'default' populator if nothing
was found.

This allows creating component styles using QML. The model given in
populate() is added in the context of the declarative item being created.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>