Add Style simple example
This commit enables the compilation of style and add it to the qml
plugin as well. The plugin will load the TestStyle unless the
COMPSTYLE environment variable is set to "qml". So to test, run

    # TestStyle
    ./run-qml.sh examples/styled-components/test.qml

    # QmlStyle
    COMPSTYLE=qml ./run-qml.sh examples/styled-components/test.qml

The QML files being used for styling are in data/qmlstyle, in this case, only
a Button.qml is there, for the dummy Button previously implemented.

This example shows a stylable (in the sense that uses
Style::populate()) C++ declarative item being used in QML, and it can be
styled either with QGraphicsWidgets from TestStyle or another QML file.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>