Support QGraphicsWidgets as components
This commit implements the two cases we have when the 'component' is a
QGraphicsWidget. When populating a QGraphicsWidget, we need to hook
the geometry of the component with the geometry of the root(s) of the
subtree(s) that were added to it.

The first case is in ComponentStyle default implementation, when the
'component' was populated by one or more QGraphicsWidgets. For this
case a layout is created and installed in the 'component', and used to
control the geometry of all direct (QGW) children of the 'component'.

The second case is in QmlStyle implementation, when 'component' was
populated by one root QDeclarativeItem object. In this case, we create
a LayoutItemWrapper class, that control the geometry of the item, and
install a layout on the 'component' to control the wrapper.

Note that both cases QGAL is used, but a simpler layout could be
implemented to do the job.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>