QmlStyle: Export component to QML context
In some cases the QML view needs more information that what is
currently available in the model. For instance, in the Button case
the view needs to access the current "text", which is a concept
that does not belong to the standard ButtonModel.

Our current approach is to expose the component to the QML world
as well as the model. Other approaches could be to either

 - Extend the model with such information (dynamic properties?)
 - Subclass the model to add that info (TextButtonModel ?)

That's something to talk about in the future.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>