Create MainApp element to handle different themes as different applications
This commit basically create a MainApp element to be used by the application's
main qml file (in this case, canola's one is main.qml). This element will be used
as the main element of the program but can be extended to have a custom settings
dialog for example.

The other change to support the change above is that the SettingsView element was
removed as it's not needed anymore for the new approach. To keep things a little
bit more organized, created a SettingsDialog.qml to keep the dialog itself and so
Settings.qml can handle it's states and effects (what the old SettingsView file did).

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>