Slideshow: removing MyRangeModel / MyButtonModel
MyRangeModel was assuming too much about its context, the connection
with knob and the action to do when the value changes should go higher
in the hierarchy.

Removed MyButtonModel as well, since the only thing it did was marking
'checkable', but this is something that the public API of each
'Button' should provide (and it does), since 'Button' can be used for
non-checkable as well.

Did the change for all the non-stylable versions.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>