Slideshow: make the interface to Settings clearer, and cleanups
This commits makes MainApp (the shared application part, that all
"themed versions" use) to assume a certain interface from the
"Settings" component (now a delegate).

The main app will extract and connect information it needs of the
Settings instead of letting internals of Settings set properties of
the main app.

One advantage of this approach is that it keeps the "glue code"
between the children of MainApp in the same place (MainApp.qml)
instead of having deep references to "viewer.state = something".  IOW,
the MainApp.qml file should give us more information in a clearrer way
about how the application works.

This commits also does some cleaning up here and there, and try harder
to use a declarative approach when it's possible to do it, e.g. the
binding of the "running" property of Viewer, which was spread in many
different lines (in ECMAScript blocks), now it's just a declarative
statement.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>