Slideshow: fix problem in Sliding animator
Just before the end of the sliding animation, sometimes the animation
"hanged". This seems to be happening because after setting the last
frame (but before painting it), the ScriptAction was executed and both
the swap of current/next image and the handler for the finished()
signal were taking too long to be processed. We think this was holding
the painting of the last frame.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>