QDeclarativeSlider: fix default populator not updating 'value'
Correct the signals in the QSlider that we use as "interface". We want
that to actually update the range model. This sounds odd, but the idea
is just reuse QSlider to demonstrate a C++ populator ;-)

Note the pattern of using setters as slots for changed signals as a
"cheap" way of doing property binding -- this happened for the
QButtonModel, too.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>