QDeclarativeButton: improve PropertyBinder utility class
Now PropertyBinder checks whether it has a signal to connect. Before
that it was connecting to index '-1' when no NOTIFY signal was
registered, which meant connecting to ALL signals. It kind of served
as a workaround, but could trigger more updates than necessary.

Added an additional constructor for allowing specification of a "fake"
NOTIFY signal, so we could use other existing signals that cover the
property change.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>