Create MX Tooltip component in QML
Look and feel is very similar to those of MX Tooltip.

The usage is a bit different though. In MX, all widgets have
intrinsic support for tooltips (set_tooltip_text method).

In QML my first approach was to make a stand alone Tooltip
component that could be added to any item, in a QML-ish
way, but for items that have internal MouseArea (PushButton for
instance) there is a conflict between that and the MouseArea
internal to the Tooltip itself. If one of the MouseAreas grab
the events, the other won't, and I don't want that.

So as a first approach, I integrated a Tooltip to the PushButton
internally, as to use only one MouseArea to control it, and
exposed a tooltip text property, similar to the one in MX.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>