First implementation of the Combobox
The combo still lacks the scrollbar when the number of items on
the model is bigger than the number of elements that the combo
can show. This will be added as soon as the scrollbar is ready.

Just like the tooltip, the combobox needs a way to show it's list
on top of any other widget displayed on the canvas. This version
still lacks this feature.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>