Mx ScrollBar: Hide states and bugfixes for vertical mode
Moving the "horizontal" and "vertical" states to an internal
item as to hide from outside world and avoid conflicts with
user-defined states. (As the "shown" state in ScrollView)

Also, fix the 'value' logic for the vertical state and leave
the Button 1/2 'source' definition to the states implementation,
to keep it consistant with the 'handle' component.

Finally, make the bar advance on button press rather than on
click, to mimic Mx behavior.

TODO: Implement auto-repeat.
TODO: Implement large-step when 'bar' is clicked

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>