mx-like: stepIncrement and pageIncrement in ScrollBar
stepIncrement is the delta [x|y] applyed to the knob position when an
arrow button is pressed. It's called singleStep in QScrollBar.
pageIncrement, in its turn,  is the delta applyed when the scrollbar path
(aka page control) receives a click. In QScrollBar, it's known as pageStep.

The names used follow the names used in MX; the default values are the ones used
in test-mx.

Signed-off-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>