Mx.Tooltip: always shows tooltip inside the window
To enable this, we use the feature of TopLevelItem, that allows it to
limit the size according to window size.

One problem that it surfaced is that the Text (Label in our case) in
the tooltip was being used to set the "preferred size" of the tooltip
itself, but to make it elide, we would need to set the Text size,
losing our information about the "preferred size" in the text.

One solution to this would be to exported the information of
"textIdealWidth" or something like that. What we did to simulate this
was: create a "model" label item, which is invisible and without eliding.

The "model" feeds the information to the tooltip preferred size, then
the "real label" (which has eliding) is anchored to the size of its
parent (which is a special one due to TopLevelItem implementation),
properly constrained by the window size.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>