Add Expander example
It still need some minor bug fixes but it's almost there. To
fix the bug of scrollbar images, created another link for now
(need proper fix on QML itself) and exported the amount of buttons
as a property of the ScrollGrid so we are able to use the correct
number for this example.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>