Mx.ButtonGroup: refresh() slot and deal with destroyed items
Now if an item is destroyed during runtime, ButtonGroup notices and
remove the connections. Also, when new items are added in the subtree
of the ButtonGroup, a 'refresh()' must be called to ensure that all
buttons are connected.

The approach to implement 'refresh()' was a brute-force disconnect all
and connect again, this can be improved.

Note that just reimplementing 'itemChange()' isn't enough, since the
items are not necessarily direct children. This issue is something to
ponder when discussing the 'ButtonModel' of qt-components.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>