mx-like: "Remove button" for the button group tab in test-mx
This adds support for the "Remove button" feature in the button group
example, which removes the rightmost button in the group.

This was tricky because it was not easy to hold a proper list of
objects in QML, so moved the code to a helper ECMAScript file. There
was easier to store the list of created buttons.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>