Temporarily removing mx-like, but don't worry
We discussed and the idea is that the master branch includes only code
that is part of "qt-components" itself. Then a 'mx' branch is going to
be created to track mx-specific stuff.

I'm removing this and then reverting inside mx branch to avoid merge
conflicts when integrating between the trees. So both branches are
starting in the same point (without mx-like) and the mx branch will
revert this commit to get the files back.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>