Integrate mx-like in the main tree
Move mx plugin and qml files no 'modules/mx' and the examples to the
main tree examples directory. Some reorganization in the examples
might still be needed.

The installation code of the plugin is still missing, but you can use
the ./run-mx.sh script to run the mx examples after compiling everything.

Note that now the mx plugin depends on the components library (since
some of the models are going to be there).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>