QRangeModel: Make it lazy to avoid issues with QML
As we can't guarantee which will be the order that QML will use
to define item's properties, it's hard to handle value's boundaries
on a set() method if it depends on another value for some reason.

This commit makes a major refactor on the code, introducing this
'lazyness' to avoid some problems.

Now d->value and d->pos always stores 'relative' values inside it while
the public methods always expects/handles 'absolute' values.

This way a user calling value()/position() will always get the right value,
which means that it will be bounded and the start value will be already
considered. Using the same logic, when using the setters - setValue()/setPosition()
- it must be always used an absolute value.

After that it was possible to do some code clean up too.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>