Remove QT_NO_SHORTCUT from compilation
The define for QT_NO_SHORTCUT was removed from compilation, and
disable the sections of code from QButtonModel that used shortcut (but
weren't compiling). It was removing functionality from a helper class
for QLineEditModel.

Shortcut support probably will be done in a different way for
QButtonModel (or even not done), so when that's decided, remove the
code sections.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>