QLineEditModel: reorganize code in the model
This commits reorganizes the code of QLineEditModel and create a
private class for it (do it now for the future's sake). The code should be
the same except for new signals emitted when some properties change.

References to 'tripleClick' and 'cancelText' were removed
completely. Some methods were renamed to avoid confusion ('end(bool
marker)' vs. 'end()'). Some new comments were added with questions :)

Some things still need to be done:

- API stills needs cleanup, specially regarding the public status of some
  methods, and which ones should be slots.
- In some property we "cheat" in some NOTIFY signals, choosing one that always
  fire when the property change (so doesn't miss property changes), but fire
  also when the property doesn't change (so false positives).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>