QLineEditModel: Remove text layout information from API
This commit removes the text layout code from the model. The current
implementation still has an internal QTextLayout, to access
functionality of QTextEngine, however it doesn't even set the font on
it. The theory here is that at some point we will not need QTextLayout
at all.

To keep our example running, another helper was created to deal with
QTextLayout and the X position of cursor and such. This is another
replication since the Text element that shows our text on the screen
also have a QTextLayout. Maybe the Text primitives should export this
kind of information? TextInput QML element does export some methods to
map character position to X-coordinate.

Note that for QML world, a TextInput is a better primitive for widget
creators, however, our point is that TextInput should use internally
our Model, which is _almost_ true since it uses QLineControl, which is
a superset of our model.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Artur Duque de Souza <artur.souza@openbossa.org>