Create MxTopLevelItem in C++
This is an implementation detail of the MX component TopLevelItem.

Some widgets like the Tooltip and Combobox need to show information
on top of any other widgets on the screen, regardless of their Z value.

To implement that we need to have an item that lives outside the item
tree where the actual tooltip/combobox/etc is, so it can have a different
Z value and stack on top of everything else.

This class is a subclass of QDeclarativeItem that traverses the item tree,
moves itself to the appropriate place and follows a "target" item, that stays
in the place the item is to be shown.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>