Mx.PathBar: functionality implemented
This commit implements the functionality available in existing Mx
PathBar widget and an example. Most of the hard work is being done at
JavaScript level. It still doesn't use the correct images and sizes
yet.

I did try to match the names for concepts in Mx, maybe this is
something we might want to tweak a little bit.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>