Use TooltipLoader in components with more than one MouseArea
The task here is similar to adding TooltipLoader to items with only
one area. The difference is that our 'shown' condition needs to
watch several items rather than only one.

Besides that, we had to do some debouncing on the 'containsMouse'
property to prevent the tooltip from being shown in the transition
between two adjacent areas.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>