Using setPosition() on Slider widget to fix a bug
When doing 'knob.x = mouseX' we were breaking the property binding
and updating the value in an external way (using a button for example)
showed that the know was not following the model's position anymore.

Now, instead of changing the knob.x position we update the position
directly on the model and then the knob follows it automatically.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>