Redone the directory structure
All source code is now in src:

src/lib		libQtComponents.so
src/module	QML module

In addition, libQtComponents is now minimal. It
should only contain the parts of qt-components
that have to be exposed to the C++ developer.
This implies that models and most other things are
in the module.

Reviewed-By: Leo