MeegoTouch StyleWrapper: Creating class to provide style info
The StyleWrapper class is meant to provide MeeGo Touch style information
to QML primitives.

Our previous approach was to create individual style classes for each
StyleContainer available in MeeGo, for instance MButtonStyleContainer or
MProgressIndicatorStyleContainer.

The single StyleWrapper class is supposed to handle all StyleContainers
thus reducing the number of boilerplate classes we have to create and
to expose to QML.

MStyleWrapper has two main properties:
 - "styleType": Defines which StyleContainer should be instantiated internally,
                for instance "Button" or "ProgressIndicator"
 - "mode": Sets the current StyleContainer mode, for instance "pressed"
           or "default"

Data is accessed from the underlaying style either by:
 - Special primitives like "ScalableImage" that can handle StyleWrappers
   directly, thus hidding non-standard MeegoTouch data-types.
 - Through additional StyleContainer properties as "fontColor"

I am not that confortable with the latter approach yet. But other alternatives
seems worse to me.

To be done:
 - Handle image updates when ThemeServer is out-of-process
 - Use enums rather than strings in the "styleType" and "mode" properties.
 - Get feedback on those new API ideas

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>