MDeclarativeScalableImage: Handle pending pixmaps issue
In MeeGo the theme server might be out-of-process. In that situation
MScalableImages may not be ready to use upon creation.

This commit makes MDeclarativeScalableImage aware of that. Whenever
we get a new image, we test its size. If it's  1 x 1 px  we assume
it is not ready yet and start listening to MTheme signals in order
to repaint the primitive when ready.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>