MDeclarativeScalableImage: Create primitive to paint MScalableImages
MScalableImages are the type of images returned by MStyles, we need
a primitive to paint them.

In order to avoid exporting that type to QML and also to reduce the
number of properties in StyleWrapper we made that primitive able to
handle StyleWrappers directly and fetch the images directly from their
underlying styles.

With that, the two properties MDeclarativeScalableImage have are:
 - style: A StyleWrapper object where it will fetch the image from
 - imageProperty: The name of the style property that holds the image,
                  that can be for instance "backgroundImage" or any
                  other image property.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>