MDeclarativeBackground: Create class to draw MWidget backgrounds
This primitive addresses two issues:
 - The need for a primitive that can draw MBackgroundTiles
 - The boilerplate logic used to decide whether MWidget backgrounds
   are drawn using backgroundImage(), backgroundColor() or backgroundTiles()

As MDeclarativeScalableImage, MDeclarativeBackground is able to fetch
the information it needs right from StyleWrapper thus hidding all
the inner magic from the QML world.

Note that since this primitive is only used to draw backgrounds, there
is no need to set which property it should read from the StyleWrapper.

Also, as MBackgroundTiles are only used in that situation, we felt no
need for a primitive that handles such class in a more generic way.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>