MDeclarativeBackground: Add TilePosition property
When using MBackgroundTiles to paint we need to know which is the
relative position of the item. This allows for instance to make
three buttons look like they merge in a group.
      __________________________
     /        |        |        \
    (  Left   | Center |  Right  )
     \________|________|________/

In Meego the value of this property is set by the current buttongroup
based on its layouting rules.

This commit also adds a proof-of-concept implementation of ButtonGroup
where items are aligned in a Row.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>