MDeclarativeIcon: Create primitive able to handle icon identifiers
In MeeGo widgets like MButton provide the setIconID API. This allows
for application developers to add an icon to these widgets by setting
a standardtized iconID which is then resolved to an actual pixmap
based on the current theme.

The name resolution in this case is handled by MTheme. To enable
QML component developers to use that infrastructure I'm creating
this class that, given an ID, searches for the correct pixmap
and draws it.

Additionally, when given a StyleWrapper, this primitive use the
"iconSize" property of the style to set its intrinsic size.

See main.qml for an usage example.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>