Yet another repository reorganization
After discussion during Leonardo visit to INDT office. The repository
now consists of

- src/lib: the library, which is minimal and right now contains only
  QDeclarativeWindow;

- src/module: Qt.labs.components module, for now contains helper
  components for designing UI elements. Next commits will clean this
  up a bit.

and two "component sets" that use more or less the Components

- src/Mx: Qt.labs.Mx module, which is a component set based on Mx
  library from Intel, implemented with QML and Qt;

- src/MeeGo: com.meego (and com.meego.themebridge) modules, which is
  the component set that uses MeeGo Touch library Theme/Style and make
  them available in QML.

The examples were moved into mx and meego directories, except the player
that we want to make it work with both sets.

To compile the "component sets" you have to add '-config mx'
and/or '-config meego'. The '-config meego' will also add MeeGo-specific
functionality to the lib and module.

Reviewed-by: Leo