MeegoTouch Switch: Create QML component
Updating MStyleWrapper and using the newly added MaskedImage
to create the Switch component.

This component looks and behaves like MeegoTouch Button when using
a SwitchType view.

As a note, we do not support the whole MButton API in this widget.
That decision was taken because much of the MButton API does not
make sense in this case. In fact, MeegoTouch itself have undefined
behavior when part of its API is used like in the hypothetical example
below:

button = new MButton()
button->setViewType(MButton::SwitchView)
button->setIconID("something");

setIconID does nothing in this case, so of course we ignore that
in the QML version.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>