Checkable and CheckableGroup components
While it is easy to create a simple "two-state switch" button component
by using QML states, for more complex setups, e.g. a "radio group" or a
set of "mutually exclusive toggle buttons", it is convenient to have
helper component to keep the invariant.

The Checkable alone also provides the "two-state switch" functionality, but
it is more useful when used together with a CheckableGroup, which provides
exclusivity for a set of Checkable items.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>