MeeGo Touch: implement ButtonRow/Column and adapt Button to work with it
This commit uses the recent added Checkable to the MeeGo Button
component, and provide a pair of components: ButtonRow and ButtonColumn.
They will pretty much emulate the common usage of MButtonGroup plus
a linear layout in a QMLish way.

One example was added to demonstrate the usage of the components.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>