Add a clear() method to CheckableGroup
This provides a direct way to clear a CheckableGroup, avoiding the
previously used way of iteratively removing each item.

This commit also updates the ButtonRow/Column in MeeGo Touch to
make use of it. This fixes the strange behavior in the gallery when
changing from One to Many in the "Button Groups". What happened there
is because of the iteration, all the items after the originally checked
ended up checked as well.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>