Meego Touch Pages: initial version
The pages navigation is currently implemented inside
the Window element, as the titlebar needs to draw a drop shadow
over the Page element. There are 2 main functions: nextPage(), which
receives a Component as parameter, and prevPage(), which pops
the currentPage from the history stack.

Doesn't have integration with titlebar content yet.

Pages dont have any customization yet, only the default behaviors are
implemented:

- Page is always shown
- Horizontal sliding animations
- Escape button will automatically be set to back or close, depending
on whether the navigation history is > 1

Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>