MDeclarativeMaskedImage: Refactory of painting to avoid X11 artifacts
We were using a QPixmap as offscreen buffer to compose the masked
image. The problem is that when running with native (X11) backend
that led to different strange behaviors in different environments.

Leo was seeing painting artifacts. I was experiencing X11 crashes :-/
By using a QImage we are safe now.

Also since I had my hands dirty, I'm now creating the masked buffer
only once to reduce the cost of each paint event.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>