New: IconButton component
This component mimics the looks and behavior of MeeGo's MButton
component when using the "icon" view type.

Should be used to paint Home, Back and other icon buttons where
shrink and glow effects are desired.

As soon as this functionality is available in MStyleWrapper,
we will be able to set a proper objectName to this component in
order to tweak the icon size and animation parameters.

For instace, the icon button that will be used to paint the Home
button should probably have a different name from a close button
that is used in a pop-up dialog.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>