MeeGo Touch: make MStyleWrapper access the MStyle directly
Instead of using MStyleContainers in our QML wrapper, we access the
MStyle directly, using MTheme::style() factory. This allows us to get
rid of the switch-statement to create StyleContainers, as well as
enable us to use future features in that factory -- like the possibility
of passing ascendant class names for the CSS (we'll send a patch for that
to libmeegotouch).

In this commit we also get rid of the enumerations, and instead export
all the key properties that one can need to specify a MStyle. While
the usage of enumeration is more "safe", it caused us to every time
make changes/additions to C++ code to "unlock" new styles. This way we
simply follow the naming available in CSS files.

Since regular users of the components are not expected to deal with
"Style {}", I think it's OK to trade "safe" for "flexibility" here.

Last, but not least, we now export the actual MStyle object to the QML,
this hopefully will make no further changes necessary to the wrapper to
use a new widget.

However this uncovers an issue in MStyle classes: their Q_PROPERTIES
do not have NOTIFY signals. While this makes sense for libmeegotouch,
since they don't change the values of a style once it is created, in QML
it causes warnings when we bind the variables.

The solution of marking the properties CONSTANT (suggested in QML
documentation) also doesn't work for MStyles since they actually use
the WRITE setters, since a MStyle is filled dynamically using
QObject::setProperty(). We'll try to fix those warnings in a later commit.

What is still missing is that MStyleContainer cached by itself the
different modes. We can easily support that, I just tried to kept this
commit smaller. Actually, we could even try to make different
MStyleWrappers share the same MStyle at a later point (while the
internal data of the styles is shared, we could cut the cost of
acquiring it by sharing the cache).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>