MeeGo Touch: use new MStyleWrapper in components
Change the components to access theme data directly from the MStyle
class instead of passing through MStyleWrapper. Note that
preferred{Width,Height} was kept in the wrapper since it does some
pre-processing of the style data to be fit to usage.

This will uncover warnings regarding binding to non-NOTIFY-able
properties from MStyle classes.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>