MeeGo Touch: fix MDeclarativeIcon
Our Icon primitive was unaware of the change of themes. Most of the
time it didn't crashed because the MThemeDaemon replaces the old
pixmap data with a new one, and when Icon paint event happened again,
the new pixmap was available for us.

We could avoid the crash just by going back to the pending state.
However when we ask for the pixmap, we ask for the default size of the
icon, which can change between themes. MThemeDaemon updates the images but
keeps the old sizes.

This issue happened with the Home/Close buttons, when changing between
themes that had different icon sizes. To fix this we release the
pixmap and ask for a new one when the theme is changed.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>