MeeGo Touch: implement cache for modes in MStyleWrapper
Implement a cache for style modes like the one in
MStyleContainer. This was the missing feature in the commit that dropped the
usage of MStyleContainer. The MStyleWrapper will keep the styles of
all "visited" modes.

While libmeegotouch does some caching of the style information, this is inside
the MTheme::style() call (which our cache now reduces the number of calls), and
the costs of getting there, even in a cache hit, are high enough to justify this
improvement.

In some cases also relieves us from a cost in acquiring/releasing the
resources (which are going to be alive anyway), like the scalable
images for the button in different modes.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>