MeeGo Touch: move MStyleWrapperManager to MThemeBridge
MStyleWrapperManager was a singleton that connected to the
themeIsChanging() signal and updated our style wrappers in correct
order: invalidating all first, and only then allowing them ask for
new data.

This commit renames this manager to MThemeBridge, since we intend to
use it to track the MDeclarativeIcon instances and ask them to
refresh(). This will avoid each Icon creating/tracking a connection.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>