MeeGo Touch: make MThemeBridge refresh the icons
Now the MThemeBridge is also responsible for refreshing the icons when
the theme changes.  This avoids a per-icon connection to the
themeIsChanging() signal that would perform the same action on each
one of them.

This commit also changes the MThemeBridge to be always connected to
the themeIsChanging() signal, instead of only connecting when someone
is registered. In practice, when MThemeBridge is constructed, it'll
be to register an icon or style wrapper anyway.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>