QDeclarativeWindow: explicitly delete the root object
Explicitly delete the root object when destroying the
QDeclarativeWindow. While we could let the QGraphicsScene take care of
deleting it (since it takes ownership), the problem is that we need to
guarantee that the QDeclarativeEngine is destroyed after the scene.

To achieve that I could simply reorder the members to guarantee that
the engine was declared first (and destroyed last), but I rather have
an explicit solution than relying on the order the members are
destroyed.

From QGraphicsScene point of view, it's OK to destroy items that are
on the scene, they'll be correctly removed. By the way, this is the
same approach done by QDeclarativeView.

This fixes the errors we were getting when using the
QDeclarativeWindow in the gallery demo: the onDestruction handlers
were called but couldn't find properties or ids of their components,
e.g. the Checkable.qml writing on its 'exclusiveGroup' property.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>