Fix error when exiting meego gallery example
There's no checkedButton when we are exiting the example because the
contents of the ButtonColumn is destroyed. So we can't assume that
checkedButton will always have a 'text' property. So we check it
before.

Food for thought: this kind of error might be an indication that we
should consider a export the value-based API like we have in Checkable
to the ButtonColumn/Row.  So Buttons could set a value (which defaults
to a binding to 'text' property), and the value of the checked button
would be available as a property in ButtonColumn/Row.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>