Added PositionIndicator component
To be used as a child of the Flickable (or any of its
subclasses). With Flickable itself, you must set it as part of the
"children" property.

Right now, the position indicator flash when the page appears, is made
using a timer. This is *not* the right way to do it. Ideally, we should
wait until the the parent Flickable is ready and the page appears (and
only the first time, as this is the current behavior in
libmeegotouch). However, there is no way in QML to wait for any item
ready/complete signal.

Modified the widgetgallery example to show the PositionIndicator
on list items, and as a sub-demo in the gallery example .

Reviewed-by: Leo
Task-number: QTCOMPONENTS-83